변경된 DB에 맞춰 백엔드 개발완, approve 조회기능 추가

+++ DDIS_Project/src/main/java/com/DDIS/shareTodo/Command/domain/aggregate/Entity/MemberShareTodo.java
@@ -15,7 +15,7 @@ public class MemberShareTodo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "member_share_todo_num")
-    private int memberShareTodoNum;
+    private Long memberShareTodoNum;
 
     @ManyToOne
     @JoinColumn(name = "share_todo_num")
